// src/components/pricing/types.tsx
/* eslint-disable @typescript-eslint/no-explicit-any */
export type PricingFields = {
  titleOne?: string;
  descOne?: string;
  photoOne?: any;
  titleTwo?: string;
  descTwo?: string;
  titleThreeImg?: any;
  titleThree?: string;
  groupClasses?: string;
  groupItem1?: string;
  groupPrice1?: string;
  groupItem2?: string;
  groupPrice2?: string;
  groupItem3?: string;
  groupPrice3?: string;
  personalClasses?: string;
  personalItem1?: string;
  personalPrice1?: string;
  personalItem2?: string;
  personalPrice2?: string;
  personalItem3?: string;
  personalPrice3?: string;
  extraPackage?: string;
  extraItem1?: string;
  extraPrice1?: string;
  extraItem2?: string;
  extraPrice2?: string;
  extraItem3?: string;
  extraPrice3?: string;
  title4?: string;
  subTitle1?: string;
  subTitleImg1?: any;
  title4des1?: string;
  subTitle2?: string;
  subTitleImg2?: any;
  title4des2?: string;
  subTitle3?: string;
  subTitleImg3?: any;
  title4des3?: string;
  button?: string;
};